Create basic_tools.py
The file will contain some functions that are useful to the main machine learning algorithms.

The function mean_normalization returns an array with the same size of the array that takes as a parameter but with every elment of x normalize, which means, that subtract the mean to each element and divide by the standard deviation.

The functions feature_scaling return an array with the same size of x but get every value scaled into aproximately a [-1, 1] range, because it divides each element by the maximum value of x.